# RFC: Update route syntax