--- a/.golangci.yml
+++ b/.golangci.yml
@@ -1,0 +1,6 @@
+linters:
+  enable-all: true
+
+issues:
+  exclude-rules:
+    - path: _test\.go