# Introduction
This RFC proposes to update the current route declaration format for named and catch-all parameters  by using `{}` and `*{} or {*}` as delimiters instead of `:` and `*`. This change aims to improve compatibility with common URL patterns and provide greater flexibility for mixing static and named route segments.

## Motivation
The current route declaration format using `:` for named parameters and `*` for catch-all parameters has some limitations in terms of readability, especially when working with certain API patterns. For instance, `/users/uid:1234/profile` is relatively common pattern in APIs, and the proposed change would make it easier to define such routes by providing more intuitive delimiters (e.g. `/users/uid:{id}/profile`).

Additionally, the current format limits the flexibility to mix static and named route segments in the same path segment. For example, consider the `route /date/{year}-{month}-{day}`. With the current format, this would have to be defined as `/date/:ymd` and require additional parsing of the `ymd` parameter to extract the year, month, and day. The proposed format with `{}` delimiters allows for more intuitive and straightforward route definitions. In the new format, each named parameter is defined explicitly and can be accessed directly without the need for additional parsing.

The new syntax also paves the way for more advanced routing features, such as incorporating regular expressions directly into route definitions (e.g `/users/{id:[0-9]+}`).

## Design
The proposed update involves changing the route declaration format for named and catch-all parameters as follows:

1. Replace `:` with `{}` as the delimiter for named parameters.
2. Replace `*` with `*{}` or `{*}` as the delimiter for catch-all parameters.

 ## Examples
Current declaration:
````
Pattern /users/:id
Pattern /*filepath
````
Updated declaration:
````
Pattern /users/{id}
Pattern /*{filepath} or /{*filepath}
````

## Direct benefits
This new design would also provide more flexible route declaration
1. Date and time: `/datetime/{year}-{month}-{day}T{hour}:{minute}:{second}`
Example: `/datetime/2023-03-26T18:30:15`
2. Geocoordinates: `/geo/{latitude},{longitude}`
Example: `/geo/37.7749,-122.4194`
3. Product size and color: `/products/{product_id}-size{size}-color{color}`
Example: `/products/1234-sizeM-colorRed`
4. User ID with domain: `/users/{user_id}@{domain}/profile`
Example: `/users/1234@example.com/profile`

## Drawbacks
- Potential increase in code complexity
- Potential impact on performance

## Conclusion
The proposed new syntax aims to enhance readability and flexibility when declaring routes. At the same time, the intention is to improve the routing experience without compromising efficiency and speed. As such, the acceptance of this proposal is contingent upon ensuring that the impact on performance remains minimal to negligible.