@tigerwill90 2023-03-28
Preliminary Benchmark
````
// BenchmarkTree_lookup-16    	16364988	        67.49 ns/op	       0 B/op	       0 allocs/op
// BenchmarkTree_lookup-16    	12335794	       103.0 ns/op	       0 B/op	       0 allocs/op
// BenchmarkTree_lookup-16    	 9970410	       120.7 ns/op	       0 B/op	       0 allocs/op
// vs
// BenchmarkTree_lookup-16    	38060128	        31.86 ns/op	       0 B/op	       0 allocs/op
// BenchmarkTree_lookup-16    	26331432	        46.13 ns/op	       0 B/op	       0 allocs/op
// BenchmarkTree_lookup-16    	23949165	        51.59 ns/op	       0 B/op	       0 allocs/op
````