@tigerwill90 2023-03-31
````
goos: linux
goarch: amd64
pkg: github.com/tigerwill90/fox
cpu: Intel(R) Core(TM) i9-9900K CPU @ 3.60GHz
                    │   old.txt    │                new.txt                │
                    │    sec/op    │   sec/op     vs base                  │
StaticAll-16          8.183µ ±  2%   8.649µ ± 1%   +5.70% (p=0.000 n=10)
Lookup-16             6.354µ ±  1%   7.017µ ± 1%  +10.42% (p=0.000 n=10)
GithubParamsAll-16    88.27n ±  4%   88.27n ± 2%        ~ (p=0.796 n=10)
StaticParallel-16     6.511n ±  2%   6.604n ± 2%        ~ (p=0.119 n=10)
CatchAll-16           31.89n ± 11%
CatchAllParallel-16   5.719n ± 14%   5.982n ± 7%        ~ (p=0.393 n=10)
geomean               132.7n         184.1n        +4.37%                ¹
¹ benchmark set differs from baseline; geomeans may not be comparable
````

Preliminary benchmark show a small overhead with the new implementation.